// frontend/src/generated/graphql.ts
export enum Priority {
  HIGH = 'HIGH',
  MEDIUM = 'MEDIUM',
  LOW = 'LOW',
}

export enum TaskStatus {
  TODO = 'TODO',
  IN_PROGRESS = 'IN_PROGRESS',
  DONE = 'DONE',
}

export interface Subject {
  __typename?: 'SubjectModel';
  id: string;
  name: string;
  color: string;
  order: number;
  units: Unit[];
}

export interface Unit {
  __typename?: 'UnitModel';
  id: string;
  name: string;
  order: number;
  subjectId: string;
}

export interface Task {
  __typename?: 'TaskModel';
  id: string;
  title: string;
  description?: string | null;
  priority: Priority;
  status: TaskStatus;
  startDate: string;
  dueDate: string;
  completedAt?: string | null;
  subjectId: string;
  unitId?: string | null;
  subject: Pick<Subject, 'id' | 'name' | 'color'>;
  unit?: Pick<Unit, 'id' | 'name'> | null;
  createdAt: string;
  updatedAt: string;
}